reorganize unit specific property models
